+++ src/main/java/com/aetherteam/aether/mixin/mixins/common/AbstractArrowMixin.java
@@ -2,6 +2,7 @@ package com.aetherteam.aether.mixin.mixins.common;
 
 import com.aetherteam.aether.attachment.AetherDataAttachments;
 import com.aetherteam.aether.attachment.AttachmentSyncable;
+import com.aetherteam.aether.attachment.PhoenixArrowAttachment;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
@@ -32,7 +33,7 @@ public abstract class AbstractArrowMixin {
         if (arrow.hasAttached(AetherDataAttachments.PHOENIX_ARROW)) {
             var attachment = arrow.getAttachedOrCreate(AetherDataAttachments.PHOENIX_ARROW);
             if (attachment.isPhoenixArrow() && !arrow.level().isClientSide()) {
-                attachment.setSynced(arrow.getId(), AttachmentSyncable.Direction.CLIENT, "setPhoenixArrow", true); // Sync Phoenix Arrow variable to client.
+                attachment.setSynced(arrow.getId(), AttachmentSyncable.Direction.CLIENT, PhoenixArrowAttachment.PHOENIX_ARROW_SYNC_KEY, true); // Sync Phoenix Arrow variable to client.
                 if (this.isInGround()) { // Spawn less particles when the arrow is in the ground.
                     if (this.inGroundTime % 5 == 0) {
                         this.spawnParticles(arrow);